git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@70 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/transform/mpadecfilter/dtsdec-0.0.1/libdts/dts_bitstream.c
@@ -38,7 +38,7 @@ void dts_bitstream_init (dts_state_t * state, uint8_t * buf, int word_mode,
     intptr_t align;
 
     align = (uintptr_t)buf & 3;
-    state->buffer_start = (uint32_t *) (buf - align);
+    state->buffer_start = (uint32_t *) ((uintptr_t)buf - align);
     state->bits_left = 0;
     state->current_word = 0;
     state->word_mode = word_mode;
